Add highlight and use rsvg-convert instead of svg2pdf

+++ rM2PDF.py
@@ -22,48 +22,50 @@ for file in os.listdir(input_dir):
 prefix =  [x[:-6] for x in note_line]
 
 for id,prefix in enumerate (prefix):
-    # get visible name in metdata
-    pdf_data  = open(input_dir+prefix+'.metadata')
-    visiable_name = pdf_data.readlines()
-    visiable_name = visiable_name[10].split(":",1)[1][2:-2].replace(" ", "_")
+    if prefix == '483e3894-f057-498a-a9cd-bb16e10255ac':
+        # get visible name in metdata
+        pdf_data  = open(input_dir+prefix+'.metadata')
+        visiable_name = pdf_data.readlines()
+        visiable_name = visiable_name[10].split(":",1)[1][2:-2].replace(" ", "_")
 
-    # convert .lines to .svg
-    script =  'python '+'./rM2svg.py -i {line} -o {svg_prefix}'
-    script = script.format (line=input_dir+note_line[id],svg_prefix=t+'temporary')
-    os.system(script)
-    # convert .svg to pdf
-    os.system('svg2pdf -o '+t+'"%(base)s.pdf" '+t+'*.svg')
-    os.system('rm '+t+'*.svg')
+        # convert .lines to .svg
+        script =  'python '+'./rM2svg.py -i {line} -o {svg_prefix} -c'
+        script = script.format (line=input_dir+note_line[id],svg_prefix=t+'temporary')
+        os.system(script)
+        # convert .svg to pdf
+        # os.system('svg2pdf -o '+t+'"%(base)s.pdf" '+t+'*.svg')
+        os.system('rsvg-convert -f pdf -o '+t+'note.pdf '+t+'*.svg')
+        os.system('rm '+t+'*.svg')
 
-    # check notebook or document
-    if os.path.isfile(input_dir+prefix+'.pdf'): # document
+        # check notebook or document
+        if os.path.isfile(input_dir+prefix+'.pdf'): # document
 
-        # combine notes to one pdf
-        os.system('convert '+t+'temporary*.pdf '+t+'note.pdf')
-        # check width and height of original pdf
-        width_column = "'NR==3{print $4}'"
-        height_column = "'NR==3{print $5}'"
-        width = os.popen('./cpdf -page-info '+input_dir+prefix+'.pdf'+' | awk {}'\
-            .format(width_column)).read()
-        height = os.popen('./cpdf -page-info '+input_dir+prefix+'.pdf'+' | awk {}'\
-            .format(height_column)).read()
+            # combine notes to one pdf
+            # os.system('convert '+t+'temporary*.pdf '+t+'note.pdf')
+            # check width and height of original pdf
+            width_column = "'NR==3{print $4}'"
+            height_column = "'NR==3{print $5}'"
+            width = os.popen('./cpdf -page-info '+input_dir+prefix+'.pdf'+' | awk {}'\
+                .format(width_column)).read()
+            height = os.popen('./cpdf -page-info '+input_dir+prefix+'.pdf'+' | awk {}'\
+                .format(height_column)).read()
 
-        # check landscape or portrait
-        if width > height: # landscape
-            size = "'"+height+' '+width+"'"
-            scalerotate = './cpdf -scale-to-fit {} '.format(size)+t+'note.pdf '+\
-            'AND -rotate 90 AND -upright '+t+'note.pdf '+'-o '+t+'note.pdf'
-            os.system(scalerotate)
-        else: # portrait
-            size = "'"+width+' '+height+"'"
-            scale = './cpdf -scale-to-fit {} '.format(size)+t+'note.pdf '+'-o '+t+'note.pdf'           
-            os.system(scale)
+            # check landscape or portrait
+            if width > height: # landscape
+                size = "'"+height+' '+width+"'"
+                scalerotate = './cpdf -scale-to-fit {} '.format(size)+t+'note.pdf '+\
+                'AND -rotate 90 AND -upright '+t+'note.pdf '+'-o '+t+'note.pdf'
+                os.system(scalerotate)
+            else: # portrait
+                size = "'"+width+' '+height+"'"
+                scale = './cpdf -scale-to-fit {} '.format(size)+t+'note.pdf '+'-o '+t+'note.pdf'           
+                os.system(scale)
 
-        # combine note and original pdf
-        os.system('./cpdf -combine-pages '+t+'note.pdf '+ input_dir+prefix+'.pdf '\
-            +'-o '+output_dir+visiable_name+'.pdf')
-        os.system('rm '+t+'temporary*.pdf '+t+'note.pdf')
+            # combine note and original pdf
+            os.system('./cpdf -combine-pages '+t+'note.pdf '+ input_dir+prefix+'.pdf '\
+                +'-o '+output_dir+visiable_name+'.pdf')
+            # os.system('rm '+t+'temporary*.pdf '+t+'note.pdf')
 
-    else: # notebook
-        os.system('convert '+t+'temporary*.pdf '+output_dir+visiable_name+'.pdf')
-        os.system('rm '+t+'temporary*.pdf')
+        else: # notebook
+            os.system('convert '+t+'temporary*.pdf '+output_dir+visiable_name+'.pdf')
+            os.system('rm '+t+'temporary*.pdf')
